solution to day  6. refactored day5 to use long

+++ src/test/java/Day5Test.java
@@ -9,13 +9,13 @@ public class Day5Test {
 
     @Test
     public void day5Part1() {
-        Assertions.assertEquals(BigInteger.valueOf(525792406), day5.part1());
+        Assertions.assertEquals(525792406, day5.part1());
     }
 
     @Test
     public void verifyFirstSeedGetsItsLowestLocationTest() {
         Day5 day5Test = new Day5("src/main/Day5Test");
-        Assertions.assertEquals(BigInteger.valueOf(82), day5Test.getLocationNumber(day5Test.extractSeeds().get(0)));
+        Assertions.assertEquals(82, day5Test.getLocationNumber(day5Test.extractSeeds().get(0)));
     }
 
     @Test
@@ -26,18 +26,18 @@ public class Day5Test {
     @Test
     public void day5Part2Test() {
         Day5 day5Test = new Day5("src/main/Day5Test");
-        Assertions.assertEquals(BigInteger.valueOf(46), day5Test.part2());
+        Assertions.assertEquals(46, day5Test.part2());
     }
 
     @Test
     public void verifyLowestLocationNumberInTest() {
         Day5 day5Test = new Day5("src/main/Day5Test");
-        Assertions.assertEquals(BigInteger.valueOf(35), day5Test.part1());
+        Assertions.assertEquals(35, day5Test.part1());
     }
 
     @Test
     public void verifyLowestLocationForRangePair() {
         Day5 day5Test = new Day5("src/main/Day5Test");
-        Assertions.assertEquals(BigInteger.valueOf(46), day5Test.getLowestLocationNumberForRange(BigInteger.valueOf(79), BigInteger.valueOf(14)));
+        Assertions.assertEquals(46, day5Test.getLowestLocationNumberForRange(79, 14));
     }
 }
